UPA/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.7.3            # via -r requirements.in
appdirs==1.4.4            # via black
appnope==0.1.2            # via ipython
async-timeout==3.0.1      # via aiohttp
attrs==20.3.0             # via aiohttp
backcall==0.2.0           # via ipython
black==20.8b1             # via -r requirements.in
certifi==2020.12.5        # via requests
chardet==3.0.4            # via aiohttp, requests
click==7.1.2              # via black, pip-tools, uvicorn
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via ipython
fastapi==0.62.0           # via -r requirements.in
h11==0.11.0               # via uvicorn
idna==2.10                # via requests, yarl
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r requirements.in
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via mpld3
kiwisolver==1.3.1         # via matplotlib
markupsafe==1.1.1         # via jinja2
matplotlib==3.3.3         # via mpld3, seaborn
mpld3==0.5.2              # via -r requirements.in
multidict==5.1.0          # via aiohttp, yarl
mypy-extensions==0.4.3    # via black
numpy==1.19.2             # via -r requirements.in, matplotlib, pandas, scipy, seaborn
orjson==3.4.5             # via -r requirements.in
pandas==1.1.4             # via -r requirements.in, seaborn
parso==0.7.1              # via jedi
pathspec==0.8.1           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==8.0.1             # via matplotlib
pip-tools==5.4.0          # via -r requirements.in
prompt-toolkit==3.0.8     # via ipython
psycopg2-binary==2.8.6    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
pydantic==1.7.3           # via fastapi
pygments==2.7.2           # via ipython
pymongo==3.11.2           # via -r requirements.in
pyparsing==2.4.7          # via matplotlib
python-dateutil==2.8.1    # via matplotlib, pandas
python-dotenv==0.15.0     # via -r requirements.in
pytz==2020.4              # via pandas
regex==2020.11.13         # via black
requests==2.25.0          # via -r requirements.in
scipy==1.5.4              # via seaborn
seaborn==0.11.0           # via -r requirements.in
six==1.15.0               # via pip-tools, python-dateutil
starlette==0.13.6         # via fastapi
toml==0.10.2              # via black
traitlets==5.0.5          # via ipython
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via aiohttp, black
urllib3==1.26.2           # via requests
uvicorn==0.12.3           # via -r requirements.in
wcwidth==0.2.5            # via prompt-toolkit
yarl==1.6.3               # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
